requirements.txt
pytest
selenium
webdriver-manager
pyyaml
pytest-html
pandas
matplotlib
seaborn
numpy
